ex: 11.7 - 11.9
Implementing flatMap in terms of compose is confusing because of the
type system, in essense compose[Any, F[A], B] is the result, which makes
sense because the first parameter can be ignored.

The equivalence of the associative laws can be seen because the flatMap
equivalence has the implementation of compose in the right side.